requirements.txt
watchdog>=4.0.0
Pillow>=10.0.0
zwoasi>=0.2.0
numpy>=1.24.0
opencv-python>=4.8.0
ttkbootstrap>=1.10.1
requests>=2.31.0
appdirs>=1.4.4
pystray>=0.19.4

# PySide6 Fluent UI (new modern interface)
PySide6>=6.6.0
PySide6-Fluent-Widgets[full]>=1.5.0
